Made LineContent into a class with inheritors LineContentSentence and LineContentAccessibilityRelation.
Justification function now keeps track of all the justification types that have been made, so we can check if an object is a justification type.

Moved the function for calculating how to display a LineContent from the Line class to the LineContent classes.

Created the TruthValueMarker class in anticipation of implementing an FDE prover, and added functionality for displaying it.